change bitmap from const slice to vec
Using a generic const slice didn't work as `StorableHint` is not
implemented for [u8; N], it's only implemented for each const slice
up to 32.

We could probably implement it ourselves, but the vec way was easier.
If the vec cloning ends up costing too much, modifying it back to a
slice shouldn't be too hard.