merge storage for normal roles and admin roles
Splitting role management between two structures is incorrect and
doesn't implement the system correctly.

On my previous implementation I failed to realise that a role and an
admin role are the same thing. A role acts as an "admin role" if it
has any association that allows them to set other roles.

Without taking in account this, creating hierarchical structures is not
possible, as an admin role cannot also act as a normal role, and thus,
only 2 levels of hierarchy can be defined.